* 'develop' of github.com:saio-fr/lily:
  Display users' groups & groups fix
  Add from/to to transfered conversation
  fix scroll issue in chats and clear logs el on render
  add operators id to transfered msg
  Bower components fix
  Bower.json morris change
  Add bower update on deploy
  Remove bower_components & compiled css/js
  Change .gitignore

+++ src/Lily/ChatBundle/Topic/OperatorTopic.php
@@ -8,7 +8,7 @@ use Ratchet\ConnectionInterface as Conn;
 
 class OperatorTopic implements TopicInterface
 {
-	protected $container;
+	  protected $container;
 	
     public function setContainer($container)
     {
@@ -49,6 +49,7 @@ class OperatorTopic implements TopicInterface
         $operator->avatar = $conn->User->getConfig()->getAvatar();
         $operator->firstname = $conn->User->getFirstname();
         $operator->lastname = $conn->User->getLastname();
+        $operator->groups = $conn->User->getGroupNames();
         // $operator->services = $conn->User->getServices();
         $operator->type = 'operator';
         $operator->messages = array();
@@ -68,6 +69,10 @@ class OperatorTopic implements TopicInterface
      */
     public function onUnSubscribe(Conn $conn, $topic, $users)
     {
+      
+        // Security check
+        if (!isset($conn->User)) { return; }
+        
     	  foreach ($users as $item) {
             if ($item->id === $conn->User->getId() && $item->type === 'operator') {
 				        $users->detach($item);
